github: Improve Pull Request template
By updating pull request template, this ensures that all newly patches
will be hopefully submitted to upstream, which help us that we dont need to
maintain these patches for ages. Also, all patches should be applied by git am.
This is important, because this follows OpenWrt main repo contribution policy
and we will know, who is the author of the patch and what it does (=
commit subject, commit description).

Also, we don't need to know on which host env, we compiled package.
If there is something host related, we can request details.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>